backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.1,<4.1
python-multipart
pandas
pymongo
motor
msal
requests
openai
certifi